+++ piargus/hierarchy.py
@@ -6,7 +6,7 @@ from typing import Mapping, Sequence, Tuple, Iterable
 
 import anytree
 
-from anytree_utils import from_indented, to_indented, from_rows, to_rows
+from .anytree_utils import from_indented, to_indented, from_rows, to_rows
 
 
 class Hierarchy:
@@ -14,16 +14,17 @@ class Hierarchy:
 
     __slots__ = "root", "indent", "filepath"
 
-    def __init__(self, tree=None, indent='@'):
+    def __init__(self, tree=None, indent='@', total_code="Total"):
         if not isinstance(tree, HierarchyNode):
-            tree = HierarchyNode(data=tree)
+            tree = HierarchyNode(total_code, data=tree)
 
         self.root = tree
         self.indent = indent
         self.filepath = None
 
     def __repr__(self):
-        return f"{self.__class__.__name__}({self.root.to_dict(), self.indent})"
+        return (f"{self.__class__.__name__}"
+                f"({self.root.to_dict()}, {self.indent}, total_code={self.total_code})")
 
     def __str__(self):
         # Use ascii, so we are safe in environments that don't use utf8
@@ -35,6 +36,14 @@ class Hierarchy:
     def __hash__(self):
         raise TypeError
 
+    @property
+    def total_code(self) -> str:
+        return self.root.code
+
+    @total_code.setter
+    def total_code(self, value):
+        self.root.code = value
+
     def get_node(self, path=".") -> "HierarchyNode":
         """Follow path to a new Node."""
         return HierarchyNode.resolver.get(self.root, path)
@@ -44,15 +53,15 @@ class Hierarchy:
         return max(map(len, codes))
 
     @classmethod
-    def from_hrc(cls, file, indent='@'):
+    def from_hrc(cls, file, indent='@', total_code="Total"):
         if isinstance(file, (str, Path)):
             with open(file) as reader:
                 hierarchy = cls.from_hrc(reader, indent)
                 hierarchy.filepath = Path(file)
                 return hierarchy
 
-        root = from_indented(file, indent=indent, node_factory=HierarchyNode, root_name="Total")
-        return Hierarchy(root)
+        root = from_indented(file, indent=indent, node_factory=HierarchyNode, root_name=total_code)
+        return cls(root, indent=indent)
 
     def to_hrc(self, file=None, length=0):
         if file is None:
@@ -68,14 +77,9 @@ class Hierarchy:
                                str_factory=lambda node: str(node.code).rjust(length))
 
     @classmethod
-    def from_rows(cls, rows: Iterable[Tuple[str, str]]):
+    def from_rows(cls, rows: Iterable[Tuple[str, str]], indent='@', total_code="Total"):
         """Construct from list of (code, parent) tuples."""
-
-        # Special-case pandas dataframe
-        if hasattr(rows, "itertuples"):
-            rows = rows.itertuples(index=False)
-
-        return cls(from_rows(rows, HierarchyNode, "Total"))
+        return cls(from_rows(rows, HierarchyNode, root_name=total_code), indent=indent)
 
     def to_rows(self) -> Iterable[Tuple[str, str]]:
         return to_rows(self.root, operator.attrgetter("code"), skip_root=True)